+++ packages/task/src/browser/task-frontend-contribution.ts
@@ -217,9 +217,9 @@ export class TaskFrontendContribution implements CommandContribution, MenuContri
                 isEnabled: () => true,
                 // tslint:disable-next-line:no-any
                 execute: (...args: any[]) => {
-                    const [source, label] = args;
+                    const [source, label, scope] = args;
                     if (source && label) {
-                        return this.taskService.run(source, label);
+                        return this.taskService.run(source, label, scope);
                     }
                     return this.quickOpenTask.open();
                 }